Remove extendConvertNoteToIssue
